Update wilKek.sh
Fixed yet another bug, when there's the name of the subject on the child's front page mentioned twice (in courses and tests) it would cause errors to pop up in the getNotes function and leave the subject's full name out of the final notes data sent to the parent. Also fixed the check for empty or not defined variables in the same function.